Add internal/ package; split public vs internal plumbing (2.0)
Introduce `ttkbootstrap/internal/` for implementation-detail modules and
separate them from the documented public surface (Workstream G).

- publisher (Channel/Subscriber/Publisher) is internal-only (just the legacy
  tk-widget repaint plumbing) → moved to `internal/publisher.py`. The
  top-level `ttkbootstrap.publisher` becomes a thin re-export shim that warns
  (DeprecationWarning, removed in 3.0).
- utility is documented public API, so it stays public — but the two
  implementation-detail helpers it carried move to `internal/utility.py`:
  `get_image_name` (used only by the style engine) and `center_on_parent`
  (a dialog helper). The public `utility.py` keeps `enable_high_dpi_awareness`
  and `scale_size`, and forwards the two moved names via module `__getattr__`
  with a DeprecationWarning (removed in 3.0).
- Rewire internal importers (style.py, window.py, dialogs/base.py,
  dialogs/datepicker.py) to the new internal paths.

Eliminating Publisher entirely (the bootstack version-stamped theme walk) is
deferred to the Workstream A / style.py engine design session.

Verified: 12 passed; theme-switch repaint + <Destroy> unsubscribe smoke test
passes; importing ttkbootstrap emits no deprecation warnings.
