+++ src/ttkbootstrap/internal/__init__.py
@@ -1,0 +1,7 @@
+"""Internal plumbing for ttkbootstrap.
+
+Modules under `ttkbootstrap.internal` are implementation details with no
+backward-compatibility guarantee. They may change or be removed in any release.
+Public code should import from the top-level `ttkbootstrap` package and its
+documented modules instead.
+"""
