+++ src/ttkbootstrap/utility.py
@@ -1,14 +1,11 @@
-"""Utility functions for ttkbootstrap.
+"""Public utility functions for ttkbootstrap.
 
-This module provides various utility functions for common tasks in
-ttkbootstrap applications, including high-DPI support, screen geometry
-calculations, and color manipulations.
+This module provides high-DPI support and size scaling helpers for
+ttkbootstrap applications.
 
 Functions:
     enable_high_dpi_awareness: Enable high-DPI scaling on Windows/Linux
     scale_size: Scale a size value for high-DPI displays
-    get_desktop_geometry: Get the screen dimensions
-    get_asset_path: Get the path to an asset file
 
 Example:
     ```python
@@ -22,6 +19,26 @@ Example:
     root.mainloop()
     ```
 """
+import warnings
+
+# Helpers that used to live here but are implementation details. They now live
+# in ttkbootstrap.internal.utility; accessing them through this module is
+# deprecated and the forwarding will be removed in 3.0.
+_MOVED_TO_INTERNAL = ("get_image_name", "center_on_parent")
+
+
+def __getattr__(name):
+    if name in _MOVED_TO_INTERNAL:
+        warnings.warn(
+            f"ttkbootstrap.utility.{name} is internal and has moved to "
+            f"ttkbootstrap.internal.utility.{name}; this forwarding will be "
+            "removed in 3.0.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        from ttkbootstrap.internal import utility as internal_utility
+        return getattr(internal_utility, name)
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
 
 
 def enable_high_dpi_awareness(root=None, scaling=None):
@@ -82,21 +99,6 @@ def enable_high_dpi_awareness(root=None, scaling=None):
     except:
         pass
 
-def get_image_name(image):
-    """Extract and return the tcl/tk image name from a PhotoImage 
-    object.
-    
-    Parameters:
-
-        image (ImageTk.PhotoImage):
-            A photoimage object.
-
-    Returns:
-
-        str:
-            The tcl/tk name of the photoimage object.
-    """
-    return image._PhotoImage__photo.name
 
 def scale_size(widget, size):
     """Scale the size based on the scaling factor of tkinter. 
@@ -124,26 +126,3 @@ def scale_size(widget, size):
         return int(size * factor)
     elif isinstance(size, tuple) or isinstance(size, list):
         return [int(x * factor) for x in size]
-
-
-def center_on_parent(win, parent=None):
-    """Center `win` on parent or over its master if not given"""
-    win.update_idletasks() # ensure geometry
-    if parent is None:
-        parent = getattr(win, 'master', None) or win # root if no parent
-
-    # parent geometry
-    parent.update_idletasks()
-    px, py = parent.winfo_rootx(), parent.winfo_rooty()
-    pw, ph = parent.winfo_width(), parent.winfo_height()
-    if pw <= 1 or ph <= 1:
-        # not yet realized, fallback to requested size
-        pw, ph = parent.winfo_reqwidth(), parent.winfo_reqheight()
-
-    # window geometry
-    ww = win.winfo_width() or win.winfo_reqwidth()
-    wh = win.winfo_height() or win.winfo_reqheight()
-
-    x = px + (pw - ww) // 2
-    y = py + (ph - wh) // 2
-    win.geometry(f"{ww}x{wh}+{x}+{y}")
